requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
djangorestframework-simplejwt==5.3.0
stripe==7.6.0
python-decouple==3.8
drf-spectacular==0.26.5
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
django-filter==23.3
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
